+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [v0.13.1]
 - `qflags` code prunning matching processing v2.
 - added `config.db_schemas.platforms_schemas.get_schema_as_dict`
+- improved `config.db_schemas.platforms_schemas` by adding `version` parameters
 
 ## [v0.13.0]
 - Implemented data processing V2.0 with new phencam QFLAGs and L2 and L3 derived products with weights.